+++ main/src/main/java/net/oneandone/stool/stage/Project.java
@@ -212,8 +212,6 @@ public abstract class Project {
             lazyMacros.addAll(stage.session.configuration.macros);
             lazyMacros.add("directory", getDirectory().getAbsolute());
             lazyMacros.add("localRepository", stage.session.localRepository().getAbsolute());
-            lazyMacros.add("svnCredentials", Separator.SPACE.join(stage.session.svnCredentials().svnArguments()));
-            lazyMacros.add("stoolSvnCredentials", stage.session.svnCredentials().stoolSvnArguments());
         }
         return lazyMacros;
     }